+++ pysat/instruments/pysat_testing.py
@@ -22,7 +22,7 @@ tags = {'': 'Regular testing data set',
         'no_download': 'simulate an instrument without download support',
         'non_strict': 'simulate an instrument without strict_time_flag',
         'user_password': 'simulates an instrument that requires a password',
-        'default_meta': 'simulates an instrument using the defualt meta'}
+        'default_meta': 'simulates an instrument using the default metadata'}
 
 # dictionary of satellite IDs, list of corresponding tags
 # a numeric string can be used in inst_id to change the number of points per day
@@ -45,7 +45,7 @@ preprocess = mm_test.preprocess
 
 def load(fnames, tag=None, inst_id=None, sim_multi_file_right=False,
          sim_multi_file_left=False, root_date=None, malformed_index=False,
-         num_samples=None, test_load_kwarg=None):
+         start_time=None, num_samples=86400, test_load_kwarg=None):
     """Load the test files.
 
     Parameters
@@ -68,8 +68,13 @@ def load(fnames, tag=None, inst_id=None, sim_multi_file_right=False,
         (default=None)
     malformed_index : boolean
         If True, time index will be non-unique and non-monotonic (default=False)
+    start_time : dt.timedelta or NoneType
+        Offset time of start time since midnight UT. If None, instrument data
+        will begin at midnight.
+        (default=None)
     num_samples : int
-        Number of samples per day
+        Maximum number of times to generate.  Data points will not go beyond the
+        current day. (default=86400)
     test_load_kwarg : any or NoneType
         Testing keyword (default=None)
 
@@ -89,11 +94,8 @@ def load(fnames, tag=None, inst_id=None, sim_multi_file_right=False,
     iperiod = mm_test.define_period()
     drange = mm_test.define_range()
 
-    if num_samples is None:
-        # Default to 1 day at a frequency of 1S
-        num_samples = 86400
-    uts, index, dates = mm_test.generate_times(fnames, num_samples,
-                                               freq='1S')
+    uts, index, dates = mm_test.generate_times(fnames, num_samples, freq='1S',
+                                               start_time=start_time)
 
     # Specify the date tag locally and determine the desired date range
     pds_offset = dt.timedelta(hours=12)
@@ -167,48 +169,13 @@ def load(fnames, tag=None, inst_id=None, sim_multi_file_right=False,
     data.index = index
     data.index.name = 'Epoch'
 
-    # Set the meta data
-    meta = pysat.Meta()
-    meta['uts'] = {'units': 's', 'long_name': 'Universal Time', 'custom': False}
-    meta['Epoch'] = {'units': 'Milliseconds since 1970-1-1',
-                     'Bin_Location': 0.5,
-                     'notes': 'UTC time at middle of geophysical measurement.',
-                     'desc': 'UTC seconds'}
-    meta['mlt'] = {'units': 'hours', 'long_name': 'Magnetic Local Time',
-                   'desc': 'Magnetic Local Time',
-                   'value_min': 0.0, 'value_max': 24.0,
-                   'notes': ''.join(['Magnetic Local Time is the solar local ',
-                                     'time of thefield line at the location ',
-                                     'where the field crosses the magnetic ',
-                                     'equator. In this case we just simulate ',
-                                     '0-24 with a consistent orbital period ',
-                                     'and an offset with SLT.']),
-                   'scale': 'linear'}
-    meta['slt'] = {'units': 'hours', 'long_name': 'Solar Local Time',
-                   'desc': 'Solar Local Time', 'value_min': 0.0,
-                   'value_max': 24.0,
-                   'notes': ''.join(['Solar Local Time is the local time ',
-                                     '(zenith angle of sun) of the given ',
-                                     'location. Overhead noon, +/- 90 is 6, ',
-                                     '18 SLT .'])}
-    meta['orbit_num'] = {'units': '', 'long_name': 'Orbit Number',
-                         'desc': 'Orbit Number', 'value_min': 0.0,
-                         'value_max': 25000.0,
-                         'notes': ''.join(['Number of orbits since the start ',
-                                           'of the mission. For this ',
-                                           'simulation we use the number of ',
-                                           '5820 second periods since the ',
-                                           'start, 2008-01-01.'])}
-    meta['longitude'] = {'units': 'degrees', 'long_name': 'Longitude'}
-    meta['latitude'] = {'units': 'degrees', 'long_name': 'Latitude'}
-    meta['altitude'] = {'units': 'km', 'long_name': 'Altitude'}
-    if tag != 'default_meta':
-        for var in data.keys():
-            if var.find('dummy') >= 0:
-                meta[var] = {'units': 'none',
-                             'notes': 'Dummy variable for testing'}
+    # Set the meta data.
+    meta = mm_test.initialize_test_meta('Epoch', data.keys())
 
-    return data, meta
+    if tag == 'default_meta':
+        return data, pysat.Meta()
+    else:
+        return data, meta
 
 
 list_files = functools.partial(mm_test.list_files, test_dates=_test_dates)