+++ pysat/tests/instrument_test_class.py
@@ -44,7 +44,7 @@ def initialize_test_inst_and_date(inst_dict):
     return test_inst, date
 
 
-class InstTestClass():
+class InstTestClass(object):
     """Provide standardized tests for pysat instrument libraries.
 
     Note: Not diretly run by pytest, but inherited through test_instruments.py
@@ -113,8 +113,7 @@ class InstTestClass():
 
     @pytest.mark.all_inst
     def test_standard_function_presence(self, inst_name):
-        """Test that each function is callable and all required functions exist.
-        """
+        """Test that each function is callable, all required functions exist."""
         module = import_module(''.join(('.', inst_name)),
                                package=self.inst_loc.__name__)